Created login button component
