+++ src/PageView/Login/Login.css
@@ -31,28 +31,29 @@
     justify-content: center;
 }
 
-.Login .loginButtons {
+.Login .LoginButtons {
     display: grid;
     justify-content: center;
     gap: 10px;
 }
 
-.Login .loginButtons > * {
+.Login .LoginButtons > * {
     text-transform: uppercase;
 }
 
-.Login .loginButtons p {
+.Login .LoginButtons p {
     margin: 0 auto;
 }
 
 .Login .signInButton {
     font-size: 16px;
     padding: 10px;
-    background-color: #00AAEF;
+    background-color: #00aaef;
     border-radius: 3rem;
     color: #ffffff;
     font-weight: bold;
     margin-top: 10px;
+    font-family: 'Play', sans-serif;
 }
 
 .Login .btn-google {
@@ -60,6 +61,7 @@
     padding: 10px;
     font-size: 16px;
     font-weight: bold;
-    color:#00AAEF;
-    border: 1px solid #00AAEF;
+    color:#00aaef;
+    border: 1px solid #00aaef;
+    font-family: 'Play', sans-serif;
 }
