+++ src/PageView/Profile/Profile.css
@@ -8,7 +8,6 @@
 .Profile .profileBox {
     background-image: url('../../common/images/profile-background-opacity-fixed.jpg');
     background-size: cover;
-    padding: 20px;
     display: grid;
     height: 35vh;
     place-content: center;
@@ -21,7 +20,6 @@
 .Profile .profileBox-photo img {
     width: 180px;
     height: 180px;
-    background-color: grey;
     border-radius: 100%;
     display: flex;
     justify-content: center;
@@ -48,9 +46,10 @@
 }
 
 .Profile .goal {
-    background-color: lightgray;
-    border-radius: 10px;
+    background-color: #00aaef;
+    border-radius: 2rem;
     margin: 0 10px;
+    padding: 10px 0;
     display: grid;
     grid-template-columns: repeat(3, 1fr);
     justify-content: center;
@@ -64,7 +63,6 @@
 .container-content:last-child {
     display: grid;
     place-content: center;
-    margin: 30px 0;
 }
 
 .Profile .durationGraph {
